Half of Task One
Created & animated the new shield and put it under player. I activated the shield with C#. Script requirement (when shield hits an enemy plane- it should not power down but be destroyed than lose life) is still not added. I did however add some code onto enemy script to help for it's destruction.